Add basic REPL implementation
I have implemented a basic REPL supporting an exit and help command.
Additionally, I wrote the help command to dynamically generate the help
message to ensure I don't need to update the commandHelp() function to
account for the new commands as I add more commands.

Overall, this commit sets the stage for expanding the REPL.