Update the Get() function to use the generic type PokeJSON
I made this change to set the stage for future updates where we'll fetch
different JSON data. If I didn't change to a generic, I'd have to
implement Get() functions for each type of new JSON data to be adequately
parsed to their corresponding struct type. This change simplifies that
process by allowing the Get() function to support them without making
changes to the Get() function.