Implement the "explore" command
This commit introduces the "explore" command, which retrieves the names of
Pokemon in a specified area. The command is formatted as follows:
explore <area-name>

To facilitate user input processing and argument parsing, a new package
named "inputparser" has been created. This package is responsible for
collecting user input using a Scanner and efficiently parsing it into
individual arguments, simplifying integration into the REPL.