Add catch command
I have added a catch command to catch a Pokemon and store it in a Pokedex.
The user writes the command as:
catch <Pokemon-name>

A random number is generated from 0 to the Pokemon's base experience. If
that number is less than 50, the Pokemon is caught and stored in the
Pokedex.

The Pokedex is a field in the config struct of type
map[string]PokemonData.

Lastly, I have organized the pokeapi package by placing the JSON data
structs in separate files and making the requests.go file easier to read.